+++ src/Controller/Admin/BannersController.php
@@ -71,10 +71,18 @@ class BannersController extends AppController {
 	}
 	
 	/**
-     * Lists banners
+     * Lists banners.
+     * 
+     * This action can use the `index_as_grid` template.
 	 * @uses MeCms\Model\Table\BannersTable::queryFromFilter()
      */
     public function index() {
+        $render = $this->request->query('render');
+        
+        if($this->Cookie->read('render.banners') === 'grid' && !$render) {
+            return $this->redirect(['?' => am($this->request->query, ['render' => 'grid'])]);
+        }
+        
 		$query = $this->Banners->find()
 			->contain(['Positions' => ['fields' => ['id', 'name']]])
 			->select(['id', 'filename', 'target', 'description', 'active', 'click_count', 'created']);
@@ -83,6 +91,16 @@ class BannersController extends AppController {
 		$this->paginate['sortWhitelist'] = ['Banners.filename', 'Positions.name', 'description', 'click_count', 'Banners.created'];
 		
 		$this->set('banners', $this->paginate($this->Banners->queryFromFilter($query, $this->request->query)));
+        
+        if($render) {
+            $this->Cookie->write('render.banners', $render);
+            
+            if($render === 'grid') {
+                $this->paginate['limit'] = $this->paginate['maxLimit'] = config('backend.photos');
+                
+                $this->render('index_as_grid');
+            }
+        }
     }
 	
 	/**
@@ -136,6 +154,18 @@ class BannersController extends AppController {
 
         $this->set(compact('banner'));
     }
+    
+    /**
+     * Downloads banner
+     * @param string $id Banner ID
+     * @uses MeCms\Controller\AppController::_download()
+     */
+    public function download($id = NULL) {
+        $banner = $this->Banners->get($id);
+        
+        return $this->_download($banner->path);
+    }
+    
     /**
      * Deletes banner
      * @param string $id Banner ID