+++ src/Controller/Admin/PhotosController.php
@@ -63,11 +63,45 @@ class PhotosController extends AppController {
 	 */
 	public function isAuthorized($user = NULL) {		
 		//Only admins and managers can delete photos
-		if($this->request->isAction('delete'))
+		if($this->request->isAction('delete')) {
 			return $this->Auth->isGroup(['admin', 'manager']);
-				
+        }
+        
 		return TRUE;
 	}
+    
+    /**
+     * Lists photos.
+     * 
+     * This action can use the `index_as_grid` template.
+	 * @uses MeCms\Model\Table\PhotosTable::queryFromFilter()
+     */
+    public function index() {
+        $render = $this->request->query('render');
+        
+        if($this->Cookie->read('render.photos') === 'grid' && !$render) {
+            return $this->redirect(['?' => am($this->request->query, ['render' => 'grid'])]);
+        }
+        
+		$query = $this->Photos->find()
+			->contain(['Albums' => ['fields' => ['id', 'title']]])
+            ->select(['id', 'album_id', 'filename', 'description', 'created']);
+		
+		$this->paginate['order'] = ['Photos.created' => 'DESC'];
+		$this->paginate['sortWhitelist'] = ['filename', 'Albums.title', 'Photos.created'];
+		
+		$this->set('photos', $this->paginate($this->Photos->queryFromFilter($query, $this->request->query)));
+        
+        if($render) {
+            $this->Cookie->write('render.photos', $render);
+            
+            if($render === 'grid') {
+                $this->paginate['limit'] = $this->paginate['maxLimit'] = config('backend.photos');
+                
+                $this->render('index_as_grid');
+            }
+        }
+    }
 	
 	/**
 	 * Uploads photos
@@ -75,9 +109,10 @@ class PhotosController extends AppController {
 	 */
 	public function upload() {
 		//If there's only one album, it automatically sets the query value
-		if(!$this->request->query('album') && count($this->viewVars['albums']) < 2)
+		if(!$this->request->query('album') && count($this->viewVars['albums']) < 2) {
 			$this->request->query['album'] = fk($this->viewVars['albums']);
-		
+        }
+        
 		$album = $this->request->query('album');
 		
 		if($album && $this->request->data('file')) {
@@ -88,11 +123,12 @@ class PhotosController extends AppController {
 			if($filename) {
 				$photo = $this->Photos->save($this->Photos->newEntity([
 					'album_id'	=> $album,
-					'filename'	=> basename($filename)
+					'filename'	=> basename($filename),
 				]));
 				
-				if($photo->id)
+				if($photo->id) {
 					$this->set('edit_url', ['action' => 'edit', $photo->id]);
+                }
 			}
 			
 			//Renders the element `backend/uploader/response`
@@ -114,11 +150,23 @@ class PhotosController extends AppController {
                 $this->Flash->success(__d('me_cms', 'The photo has been saved'));
                 return $this->redirect(['action' => 'index', $photo->album_id]);
             } 
-			else
+			else {
                 $this->Flash->error(__d('me_cms', 'The photo could not be saved'));
+            }
         }
 
         $this->set(compact('photo'));
+    }
+    
+    /**
+     * Downloads photo
+     * @param string $id Photo ID
+     * @uses MeCms\Controller\AppController::_download()
+     */
+    public function download($id = NULL) {
+        $photo = $this->Photos->get($id);
+        
+        return $this->_download($photo->path);
     }
 	
     /**
@@ -130,10 +178,12 @@ class PhotosController extends AppController {
 		
         $photo = $this->Photos->get($id);
 		
-        if($this->Photos->delete($photo))
+        if($this->Photos->delete($photo)) {
             $this->Flash->success(__d('me_cms', 'The photo has been deleted'));
-        else
+        }
+        else {
             $this->Flash->error(__d('me_cms', 'The photo could not be deleted'));
+        }
 			
         return $this->redirect(['action' => 'index', $photo->album_id]);
     }